:recycle: Remove unnecessary interfaces and rename files
